fix e2e, cleanup dependencies
